Add user choice buttons
Adds a rock, paper and scissors button to the GUI. When clicked, these
buttons will trigger a single round of Rock Paper Scissors. Round
results are not displayed to the user.
Note that the entire game no longer triggers on page load, but instead
plays a single round on button click.